requirements.txt
django~=4.2.7
python-decouple~=3.8
djangorestframework~=3.14.0
psycopg2
Pillow~=10.1.0
djangorestframework-simplejwt~=5.3.0